extracted player controls & progressbar
