+++ src/components/Pages/Player/Player.js
@@ -2,22 +2,11 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import {useParams, useHistory} from 'react-router-dom';
 import controllersImage from 'src/images/controllersImage.svg';
-
-
-function formatDuration(time) {
-  const minutes = Math.floor(time / 60);
-  let seconds = Math.floor(time % 60);
-
-  if (seconds < 10) {
-    seconds = `0` + seconds;
-  }
-
-  return `${minutes}:${seconds}`;
-}
+import PlayerControls from './PlayerControls/PlayerControls';
+import PlayerProgressbar from './PlayerProgressbar/PlayerProgressbar';
 
 
 function Player({movies}) {
-  const porgressbarRef = React.useRef();
   const videoRef = React.useRef();
   const [isLoaded, setIsLoaded] = React.useState(false);
   const [isPlaying, setIsPlaying] = React.useState(false);
@@ -28,7 +17,6 @@ function Player({movies}) {
   const history = useHistory();
 
   const currentMovie = movies.find((item) => item.id === id);
-  const progressbarValue = `${(currentTime / duration) * 100}`;
 
   const handleExit = () => history.push(`/films/${currentMovie.id}`); // button
 
@@ -49,34 +37,11 @@ function Player({movies}) {
     }
   };
 
-  const handleFullscreenButtonClick = () => videoRef.current.requestFullscreen();
-
-  const extractVideoTime = (evt) => {
-    const progressbar = porgressbarRef.current;
-    const progressbarWidth = progressbar.clientWidth;
-    const progressbarStart = progressbar.getBoundingClientRect().left;
-    const clickedPosition = evt.pageX;
-    const clickedValue = clickedPosition - progressbarStart;
-    const timePerPixel = duration / progressbarWidth;
-
-    return clickedValue * timePerPixel;
-  };
-
   const processVideoTime = (time) => {
     videoRef.current.currentTime = time;
   };
 
-  const handleTogglerMouseDown = (evt) => {
-    processVideoTime(extractVideoTime(evt));
-
-    const updateTimeOnMouseMove = (evtMove) => processVideoTime(extractVideoTime(evtMove));
-
-    document.addEventListener(`mousemove`, updateTimeOnMouseMove);
-    document.addEventListener(`mouseup`, function handleMouseUp() {
-      document.removeEventListener(`mousemove`, updateTimeOnMouseMove);
-      document.removeEventListener(`mouseup`, handleMouseUp);
-    });
-  };
+  const handleFullscreenButtonClick = () => videoRef.current.requestFullscreen();
 
   return (
     <>
@@ -98,39 +63,19 @@ function Player({movies}) {
         <button type="button" className="player__exit" onClick={handleExit}>Exit</button>
 
         <div className="player__controls">
-          <div className="player__controls-row">
-            <div className="player__time" >
-              <progress ref={porgressbarRef} className="player__progress" value={progressbarValue} max="100"></progress>
-              <div onMouseDown={handleTogglerMouseDown} className="player__toggler" style={{left: `${progressbarValue}%`}}>Toggler</div>
-            </div>
-            <div className="player__time-value">{formatDuration(duration - currentTime)}</div>
-          </div>
+          <PlayerProgressbar
+            duration={duration}
+            currentTime={currentTime}
+            processVideoTime={processVideoTime}
+          />
 
-          <div className="player__controls-row">
-            <button type="button" className="player__play" onClick={handlePlayClick}>
-              {!isPlaying && <>
-                <svg viewBox="0 0 19 19" width="19" height="19">
-                  <use xlinkHref="#play-s"></use>
-                </svg>
-                <span>Play</span>
-              </>}
-
-              {isPlaying && <>
-                <svg viewBox="0 0 14 21" width="14" height="21">
-                  <use xlinkHref="#pause"></use>
-                </svg>
-                <span>Pause</span>
-              </>}
-            </button>
-            <div className="player__name">{isLoaded ? `${currentMovie.title}` : `Loading...`}</div>
-
-            <button type="button" className="player__full-screen" onClick={handleFullscreenButtonClick}>
-              <svg viewBox="0 0 27 27" width="27" height="27">
-                <use xlinkHref="#full-screen"></use>
-              </svg>
-              <span>Full screen</span>
-            </button>
-          </div>
+          <PlayerControls
+            onPlayClick={handlePlayClick}
+            isPlaying={isPlaying}
+            isLoaded={isLoaded}
+            title={currentMovie.title}
+            onFullscreenClick={handleFullscreenButtonClick}
+          />
         </div>
       </div>
     </>
