+++ src/utils/formatDuration.js
@@ -1,0 +1,12 @@
+function formatDuration(time) {
+  const minutes = Math.floor(time / 60);
+  let seconds = Math.floor(time % 60);
+
+  if (seconds < 10) {
+    seconds = `0` + seconds;
+  }
+
+  return `${minutes}:${seconds}`;
+}
+
+export default formatDuration;
